services/transaction-service/pyproject.toml
[project]
name = "transaction-service"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "pydantic-settings>=2.6",
    "aiokafka>=0.11",
    "shared",
]

[tool.uv.sources]
shared = { path = "../../shared" }

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
